added folder for widgets and overlays
