+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:fire_fighter/game/game.dart';
+import 'package:fire_fighter/overlays/overlays.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -27,176 +28,4 @@ void main() {
       },
     ),
   );
-}
-
-class IntiailOverLayText extends StatelessWidget {
-  const IntiailOverLayText({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(
-            'Welcome to Fire Fighter',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 32,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 24),
-          Text(
-            'by @muhammadidrees',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 36),
-          Text(
-            'Press space to use the fireEngine to put out the fires',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 8),
-          Text(
-            'Keep an eye on the heat meter',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 8),
-          Text(
-            'If it gets too hot, you lose!',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 24),
-          Text(
-            "Let's see how long you can keep the window cool :)",
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 36),
-          BlinkingText(),
-        ],
-      ),
-    );
-  }
-}
-
-class GameOverOverlay extends StatelessWidget {
-  const GameOverOverlay({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(
-            'Game Over!',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 32,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 24),
-          Text(
-            'Thank you for playing the game',
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 16,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 16),
-          Text(
-            'This game was created using Flutter and Flame\n\n'
-            'for the contest Flame Jam 3.0.\n\n\n'
-            'You can get the source code from:\n\nhttps://www.github.com/muhammadidrees/fire_fighter',
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 12,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 24),
-          Text(
-            'Refresh the screen to play again :)',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 14,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-          SizedBox(height: 8),
-        ],
-      ),
-    );
-  }
-}
-
-class BlinkingText extends StatefulWidget {
-  const BlinkingText({super.key});
-
-  @override
-  State<BlinkingText> createState() => _BlinkingTextState();
-}
-
-class _BlinkingTextState extends State<BlinkingText>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _animationController;
-
-  @override
-  void initState() {
-    _animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(seconds: 1),
-    )..repeat(reverse: true);
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedBuilder(
-      animation: _animationController,
-      builder: (BuildContext context, Widget? child) {
-        return Opacity(
-          opacity: _animationController.value,
-          child: const Text(
-            'Press space to start',
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 18,
-              fontFamily: 'PressStart2P',
-            ),
-          ),
-        );
-      },
-    );
-  }
 }
