+++ lib/overlays/overlays.dart
@@ -1,0 +1,5 @@
+import 'package:fire_fighter/widgets/widgets.dart';
+import 'package:flutter/material.dart';
+
+part 'initial.dart';
+part 'game_over.dart';
