+++ lib/widgets/widgets.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+part 'blinking_animation.dart';
